Add files via upload
Data Ingest and Cleanup for Toronto Neighborhood Segmentation and Clustering
This notebook satisfies the requirements to get, transform, and groom the data that will be used for the subsequent neighborhood segmentation and clustering exercise. This is the first of the submissions for Week 3 of the IBM Applied Data Science Capstone course offered via Coursera. The assignment may be viewed at https://www.coursera.org/learn/applied-data-science-capstone/peer/I1bDq/segmenting-and-clustering-neighborhoods-in-toronto/submit.¶
Please see the comments in the code for detailed explanations of each processing step. Thanks!